made first pass at the tokens counting, need to test
