--- sdk/python/src/freewayai/prompt_template_format.py
@@ -1,43 +1,0 @@
-from .utilities import _read_prompt_from_system_and_id, _format_template_string
-from .prompt_formatted import Prompt_Formatted
-
-# FreewayAI is property of DiligentlyAI, Inc. and is used under an MIT license.
-# by Ben Parfitt, Feb 2024
-
-def _format_prompt(prompt, variables):
-
-    messages = []
-    if "messages" in prompt:
-        messages = _format_messages(prompt, prompt["messages"], variables)
-
-    if "prompt_template" in prompt:
-        msg = {
-            "content": _format_template_string(prompt, prompt["prompt_template"], variables),
-            "role": "user"
-        }
-        if len(messages):
-            messages.append(msg)
-        else:
-            messages = [msg]
-        
-    return Prompt_Formatted(messages)
-
-def _format_messages(prompt, messages, variables):
-
-    for message in messages:
-        message["content"] = _format_template_string(prompt, message["content"], variables)
-
-    return messages
-
-def read_and_format_prompt(system_id, systems_location, prompt_id, variables):
-    prompt = _read_prompt_from_system_and_id(system_id, systems_location, prompt_id)
-    return _format_prompt(prompt, variables)
-
-def main():
-    print(read_and_format_prompt('example', 'better_template_id', type="none here", description="none that I know of"), end="\n\n")
-    print(read_and_format_prompt('example', 'better_template_id2', type="none here", description="none that I know of"), end="\n\n")
-    print(read_and_format_prompt('example', 'better_template_id3', type="none here", description="none that I know of"), end="\n\n")
-    print(read_and_format_prompt('example', 'better_template_id4', type="none here", description="none that I know of"), end="\n\n")
-
-if __name__ == "__main__":
-    main()
