+++ sdk/python/src/freewayai/utilities.py
@@ -1,5 +1,12 @@
 import json
 
+def _format_messages(prompt, messages, variables):
+
+    for message in messages:
+        message["content"] = _format_template_string(prompt, message["content"], variables)
+
+    return messages
+
 def _read_prompt_from_system_and_id(system_id, systems_location, prompt_id):
     with open(f'{systems_location}/{system_id}/prompt_templates.json') as f:
         prompts = json.load(f)
